Fix broken unit test and do some renames
